extract component query to its own hook, incl. memo
