+++ src/extensions/component/ui/component.tsx
@@ -1,32 +1,11 @@
 import React from 'react';
-import { gql } from 'apollo-boost';
 import { useRouteMatch } from 'react-router-dom';
 import { ComponentProvider } from './context';
 import { TopBar } from './top-bar';
 import styles from './component.module.scss';
-import { ComponentModel } from './component-model';
 import { NavigationSlot, RouteSlot, SlotSubRouter } from '../../react-router/slot-router';
-import { useDataQuery } from '../../ui/ui/data/use-data-query';
 import { FullLoader } from '../../../to-eject/full-loader';
-
-const GET_COMPONENT = gql`
-  query Component($id: String!) {
-    workspace {
-      getComponent(id: $id) {
-        id
-        displayName
-        version
-        server {
-          env
-          url
-        }
-        compositions {
-          identifier
-        }
-      }
-    }
-  }
-`;
+import { useComponentQuery } from './use-component-query';
 
 export type ComponentProps = {
   navSlot: NavigationSlot;
@@ -42,13 +21,8 @@ export function Component({ navSlot, routeSlot, widgetSlot }: ComponentProps) {
     params: { componentId }
   } = useRouteMatch();
 
-  const { data } = useDataQuery(GET_COMPONENT, {
-    variables: { id: componentId }
-  });
-
-  if (!data) return <FullLoader />;
-
-  const component = ComponentModel.from(data.workspace.getComponent);
+  const component = useComponentQuery(componentId);
+  if (!component) return <FullLoader />;
 
   return (
     <ComponentProvider component={component}>
